Fixed various bugs found in some of the functions for directory, one which allowed a user to add a new feed to a directory belonging to another user.
Add and remove feed is now also supported in the backed, no way to add it through the interface
yet.

Response to several of the functions like, updating entry read status, new directory, add to a directory,
new feed, delete feed, now returns a JSON response. It consists of a bool flag "success" which
will be "true" if the function succeded, and "false" if it failed. There is also a message
value that will be displayed with JavaScript in the webinterface to give feedback to the user.